+++ CHANGELOG.md
@@ -5,7 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-- No changes yet.
+### Fixed
+- protobuf error: a protobuf error created with an invalid code returns a unknown
+  error again.
 
 ## [1.54.0] - 2021-06-01
 ### Added